// @tw/pulumi/workflows/schema/workflows.d.ts
export type Assign = [
  {
    [k: string]: unknown;
  },
  ...{
    [k: string]: unknown;
  }[]
];
/**
 * Run a function and return a result.
 */
export type Call =
  | ('http.get' | 'http.post' | 'http.put' | 'http.patch' | 'http.delete' | 'http.request')
  | ('sys.sleep' | 'sys.sleep_until' | 'sys.log')
  | ('events.await_callback' | 'events.create_callback_endpoint')
  | string
  | string;
/**
 * Define what step Workflows should execute next.
 */
export type Next = string | 'end' | 'continue' | 'break';
/**
 * A switch block. A selection mechanism that allows the value of an expression to control the flow of a workflow's execution. If a value matches, that condition's statement is executed.
 *
 * @minItems 1
 * @maxItems 50
 */
export type Switch = [Condition, ...Condition[]];
/**
 * An array of objects with a single step.
 *
 * @minItems 1
 * @maxItems 100000
 */
export type StepArray1 = [
  {
    [k: string]: Step;
  },
  ...{
    [k: string]: Step;
  }[]
];
// ...
export type For1 =
  | {
      [k: string]: unknown;
    }
  | {
      [k: string]: unknown;
    };
// ...
export type Parallel1 =
  | {
      [k: string]: unknown;
    }
  | {
      [k: string]: unknown;
    };
/**
 * Raise an exception.
 */
export type Raise =
  | string
  | {
      /**
       * Error code.
       */
      code?: number;
      /**
       * Error message string.
       */
      message?: string;
      [k: string]: unknown;
    };
/**
 * A try/except structure for error handling.
 */
export type Try = Step1 | StepArray4;
/**
 * Define a retry policy to retry steps that return a specific error code.
 */
export type Retry =
  | {
      /**
       * Defines which error codes will be retried. Options include ${http.default_retry_predicate}, ${http.default_retry_predicate_non_idempotent}, or a custom predicate defined as a subworkflow.
       */
      predicate?: string;
      /**
       * Maximum number of times a step will be retried, not counting the initial step execution attempt.
       */
      max_retries?: number;
      /**
       * Block that controls how retries occur.
       */
      backoff?: {
        /**
         * Delay in seconds between the initial failure and the first retry.
         */
        initial_delay?: number;
        /**
         * Maximum delay in seconds between retries.
         */
        max_delay?: number;
        /**
         * Multiplier applied to the previous delay to calculate the delay for the subsequent retry.
         */
        multiplier?: number;
      };
    }
  | string;
/**
 * An array of objects with a single step.
 *
 * @minItems 1
 * @maxItems 100000
 */
export type StepArray3 = [
  {
    [k: string]: Step;
  },
  ...{
    [k: string]: Step;
  }[]
];
/**
 * An array of objects with a single step.
 *
 * @minItems 1
 * @maxItems 100000
 */
export type StepArray4 = [
  {
    [k: string]: Step;
  },
  ...{
    [k: string]: Step;
  }[]
];
// ...
export interface Step {
  assign?: Assign;
  call?: Call;
  args?: Args;
  result?: Result;
  next?: Next;
  switch?: Switch;
  for?: For1;
  parallel?: Parallel1;
  raise?: Raise;
  try?: Try;
  retry?: Retry;
  except?: Except;
  return?: Return;
  steps?: StepArray1;
}
/**
 * Pass arguments and their values when calling a function that accepts parameters.
 */
export interface Args {
  url?: string | string;
  /**
   * The type of HTTP request method to use. Required if using call type http.request.
   */
  method?: 'GET' | 'HEAD' | 'POST' | 'PUT' | 'DELETE' | 'CONNECT' | 'OPTIONS' | 'TRACE' | 'PATCH';
  /**
   * Header fields to supply input to the API.
   */
  headers?: {
    [k: string]: unknown;
  };
  /**
   * Body fields to supply input to the API.
   */
  body?:
    | {
        [k: string]: unknown;
      }
    | null
    | string
    | string;
  /**
   * Query fields to supply input to the API.
   */
  query?:
    | {
        [k: string]: unknown;
      }
    | string;
  /**
   * Required if the API being called requires authentication.
   */
  auth?: {
    /**
     * The type of authentication.
     */
    type?: 'OIDC' | 'OAuth2';
    /**
     * Specify token scope to limit an application's access to a user's account.
     */
    scope?: unknown[] | string;
    /**
     * Specify token scope to limit an application's access to a user's account.
     */
    scopes?: unknown[] | string | string;
    /**
     * Specifies the audience for the OIDC token. By default, it's set to the same value as url; however, it should be set to your service's root URL.
     */
    audience?: string | string;
  };
  /**
   * Time in seconds. How long a request is allowed to run before throwing an exception. Default and maximum values vary by call.
   */
  timeout?: number;
  /**
   * Connector-specific parameters.
   */
  connector_params?: {
    /**
     * Time in seconds. The end-to-end duration the connector call is allowed to run for before throwing a timeout exception.
     */
    timeout?: number;
    /**
     * Polling policy.
     */
    polling_policy?: {
      /**
       * Time in seconds. Only applies to long-running operation calls.
       */
      initial_delay?: number;
      /**
       * Time in seconds. Only applies to long-running operation calls.
       */
      multiplier?: number;
      /**
       * Time in seconds. Only applies to long-running operation calls.
       */
      max_delay?: number;
      [k: string]: unknown;
    };
    /**
     * If set to True, the connector invocation call is non-blocking if the initial request to manage or update the resource succeeds
     */
    skip_polling?: boolean;
    /**
     * OAuth2 scopes to pass to the Google API.
     */
    scopes?: unknown[] | string | string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
/**
 * Assign the result from a call to this variable.
 */
export interface Result {
  [k]: string; // Triple Whale - edited
}
/**
 * Define conditional expressions for a switch block.
 */
export interface Condition {
  /**
   * An expression to switch on.
   */
  condition: string | boolean;
  steps?: StepArray1;
  assign?: Assign;
  call?: Call;
  args?: Args;
  result?: Result;
  next?: Next;
  switch?: Switch;
  for?: For;
  parallel?: Parallel;
  raise?: Raise;
  try?: Try;
  retry?: Retry;
  except?: Except;
  return?: Return;
  [k: string]: unknown;
}
/**
 * A step.
 */
export interface Step1 {
  assign?: Assign;
  call?: Call;
  args?: Args;
  result?: Result;
  next?: Next;
  switch?: Switch;
  for?: For1;
  parallel?: Parallel1;
  raise?: Raise;
  try?: Try;
  retry?: Retry;
  except?: Except;
  return?: Return;
  steps?: StepArray1;
}
/**
 * A try/except structure for error handling.
 */
export interface Except {
  /**
   * The name of a map variable that contains the error message.
   */
  as?: string;
  steps?: StepArray3;
}
/**
 * Stop a workflow's execution and return a value, variable, or expression.
 */
export interface Return {
  [k: string]: unknown;
}